requirements.txt
pandas
numpy
matplotlib
mysql-connector-python
mariadb
sqlalchemy
pymysql
requests
pycron
